+++ NEWS.md
@@ -1,5 +1,9 @@
 # corrr (development version)
 
+- `rearrange(absolute = TRUE)` works again (@jmbarbone, #167)
+
+- Extra `cor_df` classes are no longer appended to `cor_df` objects (@jmbarbone, #168)
+
 # corrr 0.4.4
 
 - Make `colpair_map()` more robust to input column names, with the exception of ".data" and ".env" (@jameslairdsmith, #131).